Added a safe interpolator. This is a point for testing as the structure is all over the place.
The new interpolator will only include neighbouring voxels
in the calculation if they meet requirements defined in
a user supplied function.

The complication is that interpolators are used at a number of
stages in this code. One point is directly in the cost function,
the other is in the PhysicalCentralDifferenceImageFunction.

There's currently no way to set the one in PhysicalCentralDifferenceImageFunction.

It probably makes sense for the cost function interpolator to be used
by PhysicalCentralDifferenceImageFunction.